container/requirements.txt
snowflake-connector-python==3.7.0
aiohttp==3.9.3
boto3==1.34.44